rename js files to jsx, update gulpfile to monitor those files
