+++ .gitignore
@@ -1,4 +1,5 @@
 *.log
+.DS_Store
 node_modules/
 sample/vendor
 sample/data.json
