+++ react_components/PaginationBoxView.jsx
@@ -1,8 +1,8 @@
+/** @jsx React.DOM */
 'use strict';
 
 var React              = require('react/addons');
-var PaginationListView = require("./PaginationListView");
-
+var PaginationListView = require("./PaginationListView.jsx");
 
 var PaginationBoxView = React.createClass({
